refactor: make existing dapp explorer legacy and rename v2 to existing name
